requirements.txt
python-telegram-bot
requests
lxml
cloudscraper
EbookLib
